+++ frontend/src/lib/services/sns-neurons.services.ts
@@ -7,6 +7,7 @@ import {
   refreshNeuron,
   removeNeuronPermissions,
   setFollowees,
+  splitNeuron as splitNeuronApi,
   stakeMaturity as stakeMaturityApi,
   startDissolving as startDissolvingApi,
   stopDissolving as stopDissolvingApi,
@@ -19,6 +20,7 @@ import {
   stakeNeuron as stakeNeuronApi,
 } from "$lib/api/sns.api";
 import { HOTKEY_PERMISSIONS } from "$lib/constants/sns-neurons.constants";
+import { snsTokenSymbolSelectedStore } from "$lib/derived/sns/sns-token-symbol-selected.store";
 import { i18n } from "$lib/stores/i18n";
 import { snsFunctionsStore } from "$lib/stores/sns-functions.store";
 import {
@@ -26,6 +28,7 @@ import {
   type ProjectNeuronStore,
 } from "$lib/stores/sns-neurons.store";
 import { toastsError } from "$lib/stores/toasts.store";
+import { transactionsFeesStore } from "$lib/stores/transaction-fees.store";
 import type { Account } from "$lib/types/account";
 import { toToastError } from "$lib/utils/error.utils";
 import { ledgerErrorToToastError } from "$lib/utils/sns-ledger.utils";
@@ -35,10 +38,14 @@ import {
   getSnsNeuronByHexId,
   getSnsNeuronIdAsHexString,
   hasAutoStakeMaturityOn,
+  isEnoughAmountToSplit,
+  nextMemo,
   subaccountToHexString,
 } from "$lib/utils/sns-neuron.utils";
+import { formatToken, numberToE8s } from "$lib/utils/token.utils";
 import { hexStringToBytes } from "$lib/utils/utils";
 import type { Identity } from "@dfinity/agent";
+import type { E8s } from "@dfinity/nns";
 import { Principal } from "@dfinity/principal";
 import {
   decodeSnsAccount,
@@ -48,11 +55,13 @@ import {
 } from "@dfinity/sns";
 import {
   arrayOfNumberToUint8Array,
+  assertNonNullish,
   fromDefinedNullable,
   fromNullable,
 } from "@dfinity/utils";
 import { get } from "svelte/store";
 import { getAuthenticatedIdentity } from "./auth.services";
+import { loadSnsAccounts } from "./sns-accounts.services";
 import {
   checkSnsNeuronBalances,
   neuronNeedsRefresh,
@@ -114,7 +123,7 @@ export const syncSnsNeurons = async (
   });
 };
 
-const loadNeurons = async ({
+export const loadNeurons = async ({
   rootCanisterId,
   certified,
 }: {
@@ -293,6 +302,82 @@ export const removeHotkey = async ({
   }
 };
 
+export const splitNeuron = async ({
+  rootCanisterId,
+  neuronId,
+  amount,
+  neuronMinimumStake,
+}: {
+  rootCanisterId: Principal;
+  neuronId: SnsNeuronId;
+  amount: number;
+  neuronMinimumStake: E8s;
+}): Promise<{ success: boolean }> => {
+  try {
+    const token = get(snsTokenSymbolSelectedStore);
+    assertNonNullish(token, "token not defined");
+
+    const transactionFee = get(transactionsFeesStore).projects[
+      rootCanisterId.toText()
+    ]?.fee;
+    assertNonNullish(transactionFee, "fee not defined");
+
+    const amountE8s = numberToE8s(amount);
+    const amountAndFee = amountE8s + transactionFee;
+
+    // minimum validation
+    if (
+      !isEnoughAmountToSplit({
+        amount: amountAndFee,
+        fee: transactionFee,
+        neuronMinimumStake,
+      })
+    ) {
+      toastsError({
+        labelKey: "error__sns.sns_amount_not_enough_stake_neuron",
+        substitutions: {
+          $minimum: formatToken({ value: neuronMinimumStake }),
+          $token: token.symbol,
+        },
+      });
+      return { success: false };
+    }
+
+    // TODO: Get identity depending on account to support HW accounts
+    const identity = await getNeuronIdentity();
+    // reload neurons (should be actual for nextMemo calculation)
+    await loadNeurons({
+      rootCanisterId,
+      certified: true,
+    });
+    const neurons = get(snsNeuronsStore)[rootCanisterId.toText()]
+      .neurons as SnsNeuron[];
+    // User can try to split and stake at the same time for a single principal id.
+    // The call would fail but client could just try again after displaying the error to the user.
+    // `memo` parameter may become optional in the future.
+    const memo = nextMemo({
+      identity,
+      neurons,
+    });
+
+    await splitNeuronApi({
+      rootCanisterId,
+      identity,
+      neuronId,
+      amount: amountAndFee,
+      memo,
+    });
+
+    return { success: true };
+  } catch (err) {
+    toastsError({
+      labelKey: "error__sns.sns_split_neuron",
+      err,
+    });
+    return { success: false };
+  }
+};
+
 export const disburse = async ({
   rootCanisterId,
   neuronId,
@@ -412,7 +497,7 @@ export const increaseStakeNeuron = async ({
   neuronId,
 }: {
   rootCanisterId: Principal;
-  amount: bigint;
+  amount: number;
   account: Account;
   neuronId: SnsNeuronId;
 }): Promise<{ success: boolean }> => {
@@ -420,14 +505,16 @@ export const increaseStakeNeuron = async ({
     // TODO: Get identity depending on account to support HW accounts
     const identity = await getAuthenticatedIdentity();
 
+    const stakeE8s = numberToE8s(amount);
     await increaseStakeNeuronApi({
       // We can cast it because we already checked that the neuron id is not undefined
       neuronId,
       rootCanisterId,
-      stakeE8s: amount,
+      stakeE8s,
       identity,
       source: decodeSnsAccount(account.identifier),
     });
+    await loadSnsAccounts({ rootCanisterId });
 
     return { success: true };
   } catch (err) {
@@ -448,20 +535,24 @@ export const stakeNeuron = async ({
   account,
 }: {
   rootCanisterId: Principal;
-  amount: bigint;
+  amount: number;
   account: Account;
 }): Promise<{ success: boolean }> => {
   try {
     // TODO: Get identity depending on account to support HW accounts
     const identity = await getAuthenticatedIdentity();
+    const stakeE8s = numberToE8s(amount);
     await stakeNeuronApi({
       controller: identity.getPrincipal(),
       rootCanisterId,
-      stakeE8s: amount,
+      stakeE8s,
       identity,
       source: decodeSnsAccount(account.identifier),
     });
-    await loadNeurons({ rootCanisterId, certified: true });
+    await Promise.all([
+      loadSnsAccounts({ rootCanisterId }),
+      loadNeurons({ rootCanisterId, certified: true }),
+    ]);
     return { success: true };
   } catch (err) {
     toastsError(