+++ frontend/src/lib/types/sns-neuron-detail.modal.ts
@@ -6,6 +6,7 @@ export type SnsNeuronModalType =
   | "follow"
   | "add-hotkey"
   | "stake-maturity"
+  | "split-neuron"
   | "auto-stake-maturity";
 
 export interface SnsNeuronModal {