+++ frontend/src/tests/mocks/transaction-fee.mock.ts
@@ -2,13 +2,15 @@ import { TokenAmount } from "@dfinity/nns";
 import type { Subscriber } from "svelte/store";
 
 export const mockSnsSelectedTransactionFeeStoreSubscribe =
-  () =>
+  (notDefined = false) =>
   (run: Subscriber<TokenAmount>): (() => void) => {
     run(
-      TokenAmount.fromE8s({
-        amount: BigInt(10_000),
-        token: { name: "Test", symbol: "TST" },
-      })
+      notDefined
+        ? undefined
+        : TokenAmount.fromE8s({
+            amount: BigInt(10_000),
+            token: { name: "Test", symbol: "TST" },
+          })
     );
     return () => undefined;
   };