+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Proposal 99659
+* Fix disburse button in SNS neurons
+* Fix rounding error in sns transactions
+
+## Proposal 98559
+* Gzip wasm for proposal payload
+* Staking sns neuron.
+* Main “Login” page.
+* Update main menu entries.
+* Dependencies upgrade.
+* Minor bugfixes and UI improvements.
+
 ## Proposal 98082
 * Enable stake maturity for HW.
 * Hide vote pane from detail page if proposal reward status has settled.